requirements.txt
python-dotenv>=1.0.0
chromadb>=0.4.24
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.1.0
openai>=1.30.0
